made taskui as export and  fixed some arguments that needed to be passed in properly
